requirements.txt
Django==5.0.2
python-dotenv==1.0.1
openai==1.12.0
psycopg2-binary==2.9.9  # For PostgreSQL/Supabase connection
django-bootstrap5==23.4
requests==2.31.0
python-decouple==3.8
whitenoise==6.6.0  # For static files
gunicorn==21.2.0  # For production deployment
asgiref==3.7.2  # Required for Django
sqlparse==0.4.4  # Required for Django
typing-extensions==4.9.0  # Required for OpenAI
tqdm==4.66.1  # Required for OpenAI
langchain==0.1.0
langchain-openai==0.0.2 